requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.3.7
python-dotenv==1.0.0
pdfminer.six==20221105
scikit-learn==1.3.2